move chatConfig into config folder and update imports
